Add email confirmation extension
This will require test fixes… gah
