+++ registrations/lib/registrations/mailer.ex
@@ -1,12 +1,41 @@
 defmodule Registrations.Mailer do
+  use Pow.Phoenix.Mailer
   use Swoosh.Mailer, otp_app: :registrations
+
   import Swoosh.Email
   import RegistrationsWeb.SharedHelpers
 
   alias RegistrationsWeb.Router
   alias RegistrationsWeb.Endpoint
 
+  require Logger
+
   @from "b@events.chromatin.ca"
+
+  @impl true
+  def cast(%{user: user, subject: subject, text: text, html: html}) do
+    %Swoosh.Email{}
+    |> to({"", user.email})
+    |> from(adventure_from())
+    |> subject(subject)
+    |> html_body(html)
+    |> text_body(text)
+  end
+
+  @impl true
+  def process(email) do
+    # An asynchronous process should be used here to prevent enumeration
+    # attacks. Synchronous e-mail delivery can reveal whether a user already
+    # exists in the system or not.
+
+    Task.start(fn ->
+      email
+      |> deliver()
+      |> log_warnings()
+    end)
+
+    :ok
+  end
 
   def send_welcome_email(email) do
     new()
@@ -164,4 +193,10 @@ defmodule Registrations.Mailer do
       _ -> "From: #{message.from_name} <#{message.from_address}>"
     end
   end
+
+  defp log_warnings({:error, reason}) do
+    Logger.warn("Mailer backend failed with: #{inspect(reason)}")
+  end
+
+  defp log_warnings({:ok, response}), do: {:ok, response}
 end
