+++ registrations/lib/registrations_web/models/user.ex
@@ -1,6 +1,8 @@
 defmodule RegistrationsWeb.User do
   use Ecto.Schema
+
   use Pow.Ecto.Schema
+  use Pow.Extension.Ecto.Schema, extensions: [PowEmailConfirmation]
 
   use RegistrationsWeb, :model
   alias Registrations.Repo
@@ -29,6 +31,12 @@ defmodule RegistrationsWeb.User do
     field(:voicepass, :string)
 
     timestamps()
+  end
+
+  def changeset(user_or_changeset, attrs) do
+    user_or_changeset
+    |> pow_changeset(attrs)
+    |> pow_extension_changeset(attrs)
   end
 
   @required_fields ~w(email password)a
