+++ registrations/lib/registrations_web/pow/messages.ex
@@ -1,5 +1,6 @@
 defmodule RegistrationsWeb.Pow.Messages do
   use Pow.Phoenix.Messages
+  use Pow.Extension.Phoenix.Messages, extensions: [PowEmailConfirmation]
 
   def user_has_been_created(_conn), do: "Your account was created"
