+++ registrations/lib/registrations_web/pow/routes.ex
@@ -1,6 +1,7 @@
 defmodule RegistrationsWeb.Pow.Routes do
   use Pow.Phoenix.Routes
 
+  # FIXME how does confirmation affect this?
   @impl true
   def after_sign_in_path(conn), do: RegistrationsWeb.Router.Helpers.user_path(conn, :edit)
 end
