+++ registrations/lib/registrations_web/router.ex
@@ -1,6 +1,7 @@
 defmodule RegistrationsWeb.Router do
   use RegistrationsWeb, :router
   use Pow.Phoenix.Router
+  use Pow.Extension.Phoenix.Router, extensions: [PowEmailConfirmation]
 
   pipeline :browser do
     plug(:accepts, ["html"])
@@ -20,6 +21,7 @@ defmodule RegistrationsWeb.Router do
     pipe_through :browser
 
     pow_routes()
+    pow_extension_routes()
   end
 
   scope "/", RegistrationsWeb do
